// SDKs/bitbucket.org/2.0/typescript/src/sdk/models/shared/paginatedfiles.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



// PaginatedFiles
/** 
 * A paginated list of commit_file objects.
**/
export class PaginatedFiles extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=next" })
  next?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=pagelen" })
  pagelen?: number;

  @SpeakeasyMetadata({ data: "json, name=previous" })
  previous?: string;

  @SpeakeasyMetadata({ data: "json, name=size" })
  size?: number;

  @SpeakeasyMetadata({ data: "json, name=values" })
  values?: Map<string, any>[];
}